NBA-477 Fixes taxes rates

+++ webapp/app/controllers/rosters_controller.rb
@@ -1,6 +1,5 @@
 class RostersController < ApplicationController
   skip_before_filter :authenticate_user!, :only => [:show, :sample_roster]
-  before_filter :check_open_market, :only => [:create, :autofill, :add_player, :remove_player]
 
   DEFAULT_BUY_IN = 1500
   DEFAULT_REMAINING_SALARY = 100000
@@ -46,6 +45,7 @@ class RostersController < ApplicationController
     if params[:market_id]
       market = Market.find(params[:market_id])
       #Eventing.report(current_user, 'createRoster', :contest_type => contest_type.name, :buy_in => contest_type.buy_in)
+      raise HttpException.new(403, "This market is closed") unless Market.find(params[:market_id]).accepting_rosters?
       in_progress_rosters = current_user.rosters.where(:state => 'in_progress')
       in_progress_rosters.first.destroy if in_progress_rosters.count > 5
       Eventing.report(current_user, 'createRoster', :buy_in => Roster::DEFAULT_BUY_IN)
@@ -87,6 +87,8 @@ class RostersController < ApplicationController
 
   def add_player
     roster = Roster.where(['owner_id = ? AND id = ?', current_user.id, params[:id]]).first
+    raise HttpException.new(403, "This market is closed") unless roster.market.accepting_rosters?
+
     player = Player.find(params[:player_id])
     price = roster.add_player(player, params[:position])
     Eventing.report(current_user, 'addPlayer', :player_id => player.id)
@@ -95,6 +97,8 @@ class RostersController < ApplicationController
 
   def remove_player
     roster = Roster.where(['owner_id = ? AND id = ?', current_user.id, params[:id]]).first
+    raise HttpException.new(403, "This market is closed") unless roster.market.accepting_rosters?
+
     player = Player.find(params[:player_id])
     Eventing.report(current_user, 'removePlayer', :player_id => player.id)
     price = roster.remove_player(player)
@@ -171,6 +175,8 @@ class RostersController < ApplicationController
 
   def autofill
     roster = current_user.rosters.where(:id => params[:id]).first
+    raise HttpException.new(403, "This market is closed") unless roster.market.accepting_rosters?
+
     roster.fill_pseudo_randomly5
     roster.swap_benched_players! if roster.remove_benched
     render_api_response roster
@@ -209,11 +215,5 @@ class RostersController < ApplicationController
   def get_contest_type(roster, contest_type_name)
     types = roster.market.contest_types
     types.find { |contest_type| contest_type.name == contest_type_name }
-  end
-
-  def check_open_market
-    if params[:market_id]
-      raise HttpException.new(403, "This market is closed") unless Market.find(params[:market_id]).accepting_rosters?
-    end
   end
 end
